extract label UI to independant surface
